requirements.txt
httpx>=0.28.0
python-dotenv==1.0.1
pillow==11.2.1

django>=4.2,<5
djangorestframework>=3.14
stripe>=7.0.0
django-allauth>=0.57.0
PyJWT>=2.8.0
cryptography>=41.0.0

# LangChain and LangSmith for AI orchestration
langchain>=0.3.25
langchain-openai>=0.3.21
langsmith>=0.3.45
